Flex Sensor Breakout Board
Eagle schematics for the Smart Glove flex sensor breakout board. This
breakout board provides a resistive voltage divider for the glove's five
flex sensors. Two-pin jumpers are attached to allow connections to be
made, one to the flex sensor as an input, and the other as an analog
output voltage to be read by the mbed's AnalogIn pins.